Add files via upload
An x y coordinate point is assigned for each bird as well as a velocity vector. The first class therefore makes it possible to perform various vector calculations. The second is mainly used to move the animals and the 3rd to move the swarm by following certain rules of cohesion or separation. There is also a main program which generates the display of all these classes in the form of birds which are represented by dots. You can choose the number of birds you want to insert by choosing an integer that you enter in the input bar. This version still has many bugs, especially in terms of cohesion.